mocks/requirements.txt
# Mock Server Dependencies
# =======================

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.20

# HTTP client for client app mock
httpx==0.26.0

# JWT handling
pyjwt==2.8.0
cffi>=1.15.0
cryptography>=41.0.0

# Type hints and validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Python 3.11+ compatibility
typing-extensions==4.9.0
